backend/requirements.txt
Flask
Flask-SQLAlchemy
flask_cors
openai
pinecone-client
langgraph
python-dotenv
python-jose
psycopg2-binary